Fix invalid error message implementation
All children of a listbox must be options, showing error message as an li tag without role=option was invalid. An alternative approach is to use another element with aria-live=polite attribute, to announce the results to the uesr.

However, consecutive queries with no results wouldn't be announced (because the message would be the same), so add seqrch query as part of the message to make it dynamic

- Add a new element with aria-live and aria-atomic attributes for a helpful message
- Remove unnecessary attribute selector for search results.
- Add styles height/margin styles for non-empty elements only.
- Add maxLength to input