Prevent default action on arrow up/down
Up/Down arrow keys work like Home/End, changing the caret position, but olding Shift with these keys selects text so it doesn't make sense to block it.

Also, the usual action of clicking the field is to close the listbox, but since here the listbox is always shown, so simply remove the selection/visual focus.